fix(profile): surface config-file-only system prompts in the panel
The reader only emitted systemPrompt when the record/env carried one,
so a config-file-only prompt showed nowhere — the profile Instructions
block keys off record/persona prompts and flat mode filtered
systemPrompt entirely. Fall back to the config-file prompt (ConfigFile
origin, read-only) when no record/env prompt shadows it, and keep
configFile-origin systemPrompt visible in the profile's flat config
panel; record/persona prompts still dedupe against Instructions.

Also splits reader.rs unit tests into a #[path]-included sibling file
(established repo pattern) to stay under the 1000-line budget.

Addresses Will's review on #1451.
